refactor: extract ExtGState parser to a shared rendering module
The ParsedExtGState struct and parse_ext_g_state_inner helper were
declared as pub(crate) in page_renderer.rs so the separation renderer
could borrow them. Hoisting both into rendering/ext_gstate.rs lets both
renderers depend on a single module without the cross-renderer leak.
