+++ src/rendering/mod.rs
@@ -30,6 +30,7 @@
 //! 3. Rasterize paths, text, and images to tiny-skia pixmap
 //! 4. Convert to output format (PNG/JPEG)
 
+pub(crate) mod ext_gstate;
 pub(crate) mod page_renderer;
 mod path_rasterizer;
 mod separation_renderer;
